move MixerOutput to common lib
