+++ libs/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(common)
 add_subdirectory(config)
 add_subdirectory(navigator)
 add_subdirectory(receiver)
