+++ libs/common/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_library(common INTERFACE)
+
+
+target_include_directories(common INTERFACE
+        include
+)
