+++ libs/mixer/include/ackermann_strategy.h
@@ -7,6 +7,7 @@
 
 #include "drivetrain_config.h"
 #include "mixer_strategy.h"
+#include "types.h"
 
 namespace MIXER {
 
@@ -28,7 +29,7 @@ namespace MIXER {
                     float base = CONFIG::WHEEL_BASE,
                     float angle = CONFIG::MAX_STEERING_ANGLE); // constructor
 
-        MixerOutput mix(float velocity, float angularVelocity) override; //mixing function
+        COMMON::MixerOutput mix(float velocity, float angularVelocity) override; //mixing function
 
         [[nodiscard]] float
         getFrontWheelSpeed(float angularVelocity, float wheelTurnRadius, float slipAngle,
