+++ libs/mixer/include/tank_steer_strategy.h
@@ -5,6 +5,7 @@
 #ifndef OSOD_MOTOR_2040_TANK_STEER_STRATEGY_H
 #define OSOD_MOTOR_2040_TANK_STEER_STRATEGY_H
 #include "mixer_strategy.h"
+#include "types.h"
 
 namespace MIXER {
 
@@ -12,7 +13,7 @@ namespace MIXER {
         // This class implements a steering strategy for a tank-like robot.
         // It calculates the motor speeds based on the given linear and angular velocities.
     public:
-        MixerOutput mix(float velocity, float angularVelocity) override;
+        COMMON::MixerOutput mix(float velocity, float angularVelocity) override;
     };
 }
