+++ libs/mixer/src/tank_steer_strategy.cpp
@@ -3,10 +3,11 @@
 //
 
 #include "tank_steer_strategy.h"
+#include "types.h"
 
 using namespace MIXER;
 
-MixerOutput TankSteerStrategy::mix(float velocity, float angularVelocity) {
+COMMON::MixerOutput TankSteerStrategy::mix(float velocity, float angularVelocity) {
     float left = velocity - angularVelocity;
     float right = velocity + angularVelocity;
     float abs_left = left >= 0 ? left : -left;
@@ -21,7 +22,7 @@ MixerOutput TankSteerStrategy::mix(float velocity, float angularVelocity) {
     float scaled_left = (left * speed_factor);
     float scaled_right = (right * speed_factor) * -1;
 
-    MixerOutput result = {};
+    COMMON::MixerOutput result = {};
     result.speeds = {scaled_left, scaled_right, scaled_left, scaled_right};
     result.angles = {0, 0};
