+++ libs/statemanager/include/statemanager.h
@@ -5,6 +5,7 @@
 #ifndef OSOD_MOTOR_2040_STATEMANAGER_H
 #define OSOD_MOTOR_2040_STATEMANAGER_H
 
+#include "types.h"
 #include "receiver.h"
 #include "state_estimator.h"
 #include "stoker.h"
@@ -38,7 +39,7 @@ namespace STATEMANAGER {
         // max speed factor - scale the speed of the motors down to this value
         static constexpr float SPEED_EXTENT = 1.0f;
 
-        void setSpeeds(MIXER::MixerOutput motorSpeeds) const;
+        void setSpeeds(COMMON::MixerOutput motorSpeeds) const;
     };
 
 } // StateManager
