+++ libs/statemanager/src/statemanager.cpp
@@ -36,28 +36,28 @@ namespace STATEMANAGER {
         //printf("Velocity: %f ", requestedState.velocity);
         //printf("Angular velocity: %f ", requestedState.angularVelocity);
         //printf("\n");
-        MIXER::MixerOutput ackermannOutput = mixerStrategy->mix(requestedState.velocity, requestedState.angularVelocity);
+        COMMON::MixerOutput ackermannOutput = mixerStrategy->mix(requestedState.velocity, requestedState.angularVelocity);
         setSpeeds(ackermannOutput);
     }
 
-    void StateManager::setSpeeds(MIXER::MixerOutput ackermannOutput) const {
-        stokers.FRONT_LEFT->set_speed(ackermannOutput.speeds.frontLeft);
-        stokers.FRONT_RIGHT->set_speed(ackermannOutput.speeds.frontRight);
-        stokers.REAR_LEFT->set_speed(ackermannOutput.speeds.rearLeft);
-        stokers.REAR_RIGHT->set_speed(ackermannOutput.speeds.rearRight);
-        if (std::fabs(ackermannOutput.speeds.frontLeft) > 0.05) {
+    void StateManager::setSpeeds(COMMON::MixerOutput motorSpeeds) const {
+        stokers.FRONT_LEFT->set_speed(motorSpeeds.speeds.frontLeft);
+        stokers.FRONT_RIGHT->set_speed(motorSpeeds.speeds.frontRight);
+        stokers.REAR_LEFT->set_speed(motorSpeeds.speeds.rearLeft);
+        stokers.REAR_RIGHT->set_speed(motorSpeeds.speeds.rearRight);
+        if (std::fabs(motorSpeeds.speeds.frontLeft) > 0.05) {
             if (not(steering_servos.left->is_enabled())){
                 steering_servos.left->enable();
             }
-            steering_servos.left->value(ackermannOutput.angles.left);
+            steering_servos.left->value(motorSpeeds.angles.left);
         } else {
             steering_servos.left->disable();
         }
-        if (std::fabs(ackermannOutput.speeds.frontRight) > 0.05) {
+        if (std::fabs(motorSpeeds.speeds.frontRight) > 0.05) {
             if (not(steering_servos.right->is_enabled())){
                 steering_servos.right->enable();
             }
-            steering_servos.right->value(ackermannOutput.angles.right);
+            steering_servos.right->value(motorSpeeds.angles.right);
         } else {
             steering_servos.right->disable();
         }
